requirements.txt
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
Django==2.0.2
django-filter==1.1.0
django-webpack-loader==0.6.0
djangorestframework==3.7.7
httpie==0.9.9
idna==2.6
Markdown==2.6.11
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
Pygments==2.2.0
pytz==2018.3
requests==2.18.4
urllib3==1.22
